move GraphException to jit.py
same place where GraphRunner is defined
